+++ Android TFG/app/src/androidTest/java/trabajo/fin/grado/dictionary/DictionaryBasedValueGenerator.java
@@ -1,6 +1,5 @@
 package trabajo.fin.grado.dictionary;
 
-
 import net.sf.extjwnl.JWNLException;
 import net.sf.extjwnl.data.IndexWord;
 import net.sf.extjwnl.data.POS;
@@ -10,7 +9,6 @@ import java.util.List;
 import java.util.Random;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
 
 public class DictionaryBasedValueGenerator {
     public static final int DEFAULT_NUMBER_OF_WORDS=3;
@@ -26,15 +24,13 @@ public class DictionaryBasedValueGenerator {
 
     public DictionaryBasedValueGenerator()
     {
-        this(DEFAULT_NUMBER_OF_WORDS,DEFAULT_NUMBER_OF_WORDS, new Random().nextInt());
+        this(DEFAULT_NUMBER_OF_WORDS,DEFAULT_NUMBER_OF_WORDS, new Random().nextLong());
     }
-    public DictionaryBasedValueGenerator(int words, Integer seed) {
+    public DictionaryBasedValueGenerator(int words, Long seed) {
         this(words,words, seed);
     }    
-    
-    
-    
-    public DictionaryBasedValueGenerator(int minWords, int maxWords, Integer seed) {
+
+    public DictionaryBasedValueGenerator(int minWords, int maxWords, Long seed) {
         this.minWords=minWords;
         this.maxWords = maxWords;
         this.randomGenerator = new Random(seed);
